Added feature to retry parsing the session info YAML a little later if it fails.
Due to poor synchronization mechanism for session info YAML updates it's entirely possible that when starting up, IRSDKSharper could be trying to parse the session info YAML while the iRacing simulator is updating it at the same time.  If it detects a parse failure this feature will wait exactly a half second then try again.